removed collection module
