+++ packages/earth-map/src/components/collection/collection-delete/CollectionDelete.tsx
@@ -22,7 +22,7 @@ import { useTranslation } from 'react-i18next';
 
 import { Modal } from '@marapp/earth-shared';
 
-import { ICollection } from '../../../modules/collections/model';
+import { ICollection } from '../../../fetchers/locations/queries';
 import { EPanels } from '../../../modules/sidebar/model';
 import PlacesService from '../../../services/PlacesService';
 
@@ -80,9 +80,6 @@ export function CollectionDelete(props: IProps) {
       });
       // TODO: move all the following side effects under the `EARTH` action
       // also find out why actions only work with dispatch()
-      dispatch({
-        type: 'COLLECTIONS/resetCollection',
-      });
       dispatch({
         type: 'PLACES/setPlacesSearch',
         payload: {
