+++ packages/earth-map/src/components/sidebar/component.tsx
@@ -44,7 +44,6 @@ interface ISidebarPanel {
   setIndexesSelected?: (i: string) => void;
   resetMap?: () => void;
   resetPlace?: () => void;
-  resetCollection?: () => void;
   resetLayers?: () => void;
   selectedOpen?: boolean;
 }
@@ -56,10 +55,9 @@ class Sidebar extends React.Component<ISidebarPanel> {
   };
 
   public resetMap = () => {
-    const { setPlacesSearch, resetMap, resetPlace, resetCollection, resetLayers } = this.props;
+    const { setPlacesSearch, resetMap, resetPlace, resetLayers } = this.props;
 
     resetPlace();
-    resetCollection();
     setPlacesSearch({ search: '' });
     resetLayers();
     resetMap();
